[level 1] Title: 정수 내림차순으로 배치하기, Time: 4.73 ms, Memory: 31.7 MB -BaekjoonHub

+++ 프로그래머스/1/12933. 정수 내림차순으로 배치하기/정수 내림차순으로 배치하기.cs
@@ -7,7 +7,7 @@ public class Solution {
 
             for(int i = 0; i < ints.Length; i++)
             {
-                ints[i] = (int)input[i] - 48;
+                ints[i] = input[i] - 48;
             }
             Array.Sort(ints);
             Array.Reverse(ints);
